% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adulto_equivalente.R
\docType{data}
\name{adulto_equivalente}
\alias{adulto_equivalente}
\title{Tabla de valores de adulto equivalente según sexo y edad}
\format{A data frame with 222 rows and 3 variables

\describe{
\item{CH04}{sexo}
\item{CH06}{edad}
\item{adequi}{unidades de adulto equivalente}
...
}}
\usage{
adulto_equivalente
}
\description{
Tabla de valores de adulto equivalente según sexo y edad
}
\keyword{datasets}
