% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_errors.R
\name{calculate_error}
\alias{calculate_error}
\title{Calculo del desvío estándar y el coeficiente de variación}
\usage{
calculate_error(
  value,
  codigo_aglo = "Total",
  periodo_eph = "2014.03",
  measure = "cv"
)
}
\arguments{
\item{value}{Valor o vector de valores de las estimaciones puntuales para las
que se desea hallar el desvío estándar o el coeficiente de variación.}

\item{codigo_aglo}{default = "Total". String con el código numerico de aglomerado que
corresponde a la estimación o con "Total" para el total de 31 aglomerados urbanos (ver \code{\link{errores_muestrales}})}

\item{periodo_eph}{default = "2014.03". String con "2014.03" para obtener los errores muestrales de las EPH realizadas
del tercer trimestre de 2014 en adelante o con "2003.03_2014.02" para las EPH realizadas entre el tercer trimestre del 2003
y el segundo trimestre del 2014. (ver \code{\link{errores_muestrales}})}

\item{measure}{default = "cv". String con "cv" para obtener el coeficiente de variación
correspondiente a las estimaciones o con "ds" para obtener el desvío estándar.}
}
\description{
Asigna a una estimación de un total poblacional el desvío estándar o el coeficiente de variación
correspondiente en base a las tablas de errores muestrales de INDEC para EPH continua a partir del segundo trimestre 2003.
}
\details{
disclaimer: El script no es un producto oficial de INDEC.
}
\examples{

tabla <- eph::toybase_individual_2016_03 \%>\%
  eph::organize_labels() \%>\%
  dplyr::filter(AGLOMERADO == 32) \%>\%
  calculate_tabulates(x = "CH03",
                         weights = "PONDERA",
                         add.totals = "row")
tabla \%>\%
  dplyr::mutate(ds = calculate_error(Freq, codigo_aglo = "32",
   periodo_eph = "2014.03", measure = "ds"))

}
