% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canastas.R
\docType{data}
\name{canastas}
\alias{canastas}
\title{Canastas básicas alimentarias y Totales disponibles desde enero-2015 para GBA}
\format{A data frame with 37 rows and 5 variables


\describe{
\item{anio}{}
\item{mes}{}
\item{cba}{canasta basica alimentaria}
\item{ice}{inversa del coeficiente de engels}
\item{cbt}{canasta basica total}
...
}}
\usage{
canastas
}
\description{
Canastas básicas alimentarias y Totales disponibles desde enero-2015 para GBA
}
\keyword{datasets}
