% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canastas_reg_example.R
\docType{data}
\name{canastas_reg_example}
\alias{canastas_reg_example}
\title{Canastas Basicas Alimentarias y Canastas Basicas Totales segun region y trimestre}
\format{
A data frame with 12 rows and 5 variables

\describe{
\item{region}{Nombre de region}
\item{periodo}{Año.trimestre}
\item{CBA}{canasta basica alimentaria}
\item{CBT}{inversa del coeficiente de engels}
\item{codigo}{codigo de region}
...
}
}
\usage{
canastas_reg_example
}
\description{
Esta base tiene el formato necesario para el calculo de pobreza
}
\keyword{datasets}
