% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centroides_aglomerados.R
\docType{data}
\name{centroides_aglomerados}
\alias{centroides_aglomerados}
\title{centroides_aglomerados}
\format{
A data frame with 32 rows and 4 variables:
\describe{
\item{AGLOMERADO}{(Factor) Codigo de aglomerado EPH}
\item{nombre_aglomerado}{(Factor) Nombre del aglomerado EPH}
\item{lat}{Latitud del aglomerado EPH}
\item{lon}{Longitud del aglomerado EPH}
}
}
\usage{
centroides_aglomerados
}
\description{
Base con centroides de los aglomerados de EPH.
Basado en https://www.indec.gob.ar/indec/web/Nivel4-Tema-1-39-120
}
\keyword{datasets}
