% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNO.R
\docType{data}
\name{CNO}
\alias{CNO}
\title{Categorias de las 4 dimensiones del Clasificador Nacional de Ocupaciones 2001.}
\format{
Un data frame con 63 filas y 4 variables

\describe{
\item{digit}{digito para identificar la dimension del CNO}
\item{value}{valor que toma cada categoria dentro de una dimension}
\item{label}{nombre de la categoria propia a una dimension}
\item{variable}{nombre de la dimension del CNO}

}
}
\usage{
CNO
}
\description{
Categorias de las 4 dimensiones del Clasificador Nacional de Ocupaciones 2001.
}
\keyword{datasets}
