% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Creacion de Paneles.R
\name{Crea_Panel_Continua}
\alias{Crea_Panel_Continua}
\title{Pool de Datos en Panel Base Individudal EPH Continua}
\usage{
Crea_Panel_Continua(Bases, Variables, Ventana = "Anual")
}
\arguments{
\item{Bases}{Lista de bases de microdatos a utilizar para armar el pool de datos}

\item{Variables}{Vector con nombres de las variables de interés}

\item{Ventana}{Que tipo de distancia temporal entre las obseraciones. Anual o Trimestral}
}
\value{
Devuelve el pool de datos de panel y un conteo de casos consistentes
}
\description{
Permite armar un pool de datos en panel de la EPH continua a partir
de especificar una serie de bases, variables y el largo de la ventana de observación
}
\examples{
Base_1t_2018 <- descarga_base_eph(anio = 2018,trimestre = 1)[[2]]
Base_2t_2018 <- descarga_base_eph(anio = 2018,trimestre = 2)[[2]]

Lista_Bases <- list(Base_1t_2018,Base_2t_2018)
Crea_Panel_Continua(Bases = Lista_Bases,
                    Variables =c("P21","ESTADO"),
                    Ventana = "Trimestral")
}
