% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diccionario_aglomerados.R
\docType{data}
\name{diccionario_aglomerados}
\alias{diccionario_aglomerados}
\title{Diccionario de aglomerados segun diseno de registro de la EPH}
\format{
Un data frame con 32 filas y 2 variables

\describe{
\item{codigo}{Codigo de aglomerado}
\item{aglo}{Nombre del aglomerado}

}
}
\usage{
diccionario_aglomerados
}
\description{
Diccionario de aglomerados segun diseno de registro de la EPH
}
\keyword{datasets}
