% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diccionario_regiones.R
\docType{data}
\name{diccionario_regiones}
\alias{diccionario_regiones}
\title{Diccionario de aglomerados según diseño de registro de la EPH}
\format{A data frame with 6 rows and 2 variables

\describe{
\item{codigo}{código de región}
\item{region}{Nombre de región}
...
}}
\usage{
diccionario_regiones
}
\description{
Diccionario de aglomerados según diseño de registro de la EPH
}
\keyword{datasets}
