% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bases_eph0307.r
\name{get_bases_eph0307}
\alias{get_bases_eph0307}
\title{Descarga de Bases de EPH}
\usage{
get_bases_eph0307(anio = 2003, trimestre = 3, etiqueta = FALSE)
}
\arguments{
\item{anio:}{un integer entre 2016:2018}

\item{trimestre:}{un integer con el numero de trimestre: 1,2,3,4}

\item{etiqueta:}{TRUE/FALSE, opcion para etiquetar los datos}
}
\description{
Función que descarga bases de la Encuesta Permanente de Hogares del INDEC a partir de 2016
}
\details{
disclaimer: El script no es un producto oficial de INDEC.
}
\examples{

base <- get_bases_eph0307(anio = 2018, trimestre = 1, etiqueta = FALSE)
base_individual <- base[['base_individual']]
base_hogar <- base[['base_hogar']]

}
