% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_canastas.r
\name{get_canastas}
\alias{get_canastas}
\title{Descarga de canasta basica alimentaria y canasta basica total}
\usage{
get_canastas(update = F)
}
\arguments{
\item{update:}{TRUE/FALSE, opcion actualizar las canastas con los ultimos informes publicados (lleva tiempo)}
}
\description{
Función que descarga la CBA y CBT a partir de 2016
}
\details{
disclaimer: El script no es un producto oficial de INDEC.
}
\examples{

canasta <- get_canastas(update=FALSE)

}
