% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_microdata_internal.R
\name{get_microdata_internal}
\alias{get_microdata_internal}
\title{Descarga de Bases de EPH}
\usage{
get_microdata_internal(
  year = 2018,
  trimester = NA,
  wave = NA,
  type = "individual",
  vars = "all"
)
}
\arguments{
\item{year}{un integer a partir de 2003}

\item{trimester}{un integer con el numero de trimester: 1,2,3,4, para la EPH continua}

\item{wave}{un integer con el numero de onda, 1 o 2, para la EPH puntual}

\item{type}{tipo de base a descargar: 'individual' ; 'hogar', default individual}

\item{vars}{variables a seleccionar. Default trae todas las variables}
}
\description{
Funcion que descarga bases de la Encuesta Permanente de Hogares del INDEC a partir de 1996
}
