% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ocupations_clasified.R
\name{get_ocupations_clasified}
\alias{get_ocupations_clasified}
\title{Clasificador de Ocupaciones}
\usage{
get_ocupations_clasified(base)
}
\arguments{
\item{base}{Base individual de uno o más períodos}
}
\description{
Función para clasificar las ocupaciones según las 4 dimensiones del Clasificador Nacional de Ocupaciones (CNO 2001)
}
\details{
#'disclaimer: El script no es un producto oficial de INDEC.
}
\examples{

#'
bases <- dplyr::bind_rows(toybase_individual_2016_03,toybase_individual_2016_04)
bases_clasif <- get_ocupations_clasified(base = bases)

}
