% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_in_github.R
\name{is_in_github}
\alias{is_in_github}
\title{Check if file is in github repo}
\usage{
is_in_github(year = 2018, trimester = NA, type = "individual")
}
\arguments{
\item{year}{el parametro de get_microdata_internal}

\item{trimester}{el parametro de get_microdata_internal}

\item{type}{el parametro de get_microdata_internal}
}
\description{
Check if file is in github repo
}
