% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organize_cno.R
\name{organize_cno}
\alias{organize_cno}
\title{Clasificador de Ocupaciones}
\usage{
organize_cno(base)
}
\arguments{
\item{base}{Base individual de uno o mas periodos}
}
\description{
Funcion para clasificar las ocupaciones segun las 4 dimensiones del Clasificador Nacional de Ocupaciones
(CNO 2001)
}
\details{
Disclaimer: El script no es un producto oficial de INDEC.

Importante: El usuario debe garantizarse que el clasificador CNO 2001 es compatible con la base que esta utilizando
}
\examples{


bases <- dplyr::bind_rows(toybase_individual_2016_03,toybase_individual_2016_04)
bases_clasif <- organize_cno(base = bases)

}
