% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organize_labels.R
\name{organize_labels}
\alias{organize_labels}
\title{Función para etiquetar las bases de la Encuesta Permanente de Hogares.}
\usage{
organize_labels(df, base = "individual")
}
\arguments{
\item{df}{base de microdatos de la EPH}

\item{base}{string. aclaracion de si se quiere etiquetar una base 'individual' u 'hogar'}
}
\description{
Función para etiquetar las bases de la Encuesta Permanente de Hogares.
}
\details{
disclaimer: El script no es un producto oficial de INDEC.
}
\examples{
df <- get_microdata(year = 2018, trimester = 1, labels = FALSE)
df <- organize_labels(df$base_hogar, base='hogar')
}
