% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organize_ocupations.R
\name{organize_ocupations}
\alias{organize_ocupations}
\title{Clasificador de Ocupaciones}
\usage{
organize_ocupations(base)
}
\arguments{
\item{base}{Base individual de uno o más períodos}
}
\description{
Función para clasificar las ocupaciones según las 4 dimensiones del Clasificador Nacional de Ocupaciones
(CNO 2001)
}
\details{
#'disclaimer: El script no es un producto oficial de INDEC.
}
\examples{

#'
bases <- dplyr::bind_rows(toybase_individual_2016_03,toybase_individual_2016_04)
bases_clasif <- organize_ocupations(base = bases)

}
