% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organize_panels.R
\name{organize_panels}
\alias{organize_panels}
\title{Pool de Datos en Panel - Base Individudal EPH continua}
\usage{
organize_panels(bases, variables, window = "anual")
}
\arguments{
\item{bases}{Lista de bases de microdatos a utilizar para armar el pool de datos}

\item{variables}{Vector con nombres de las variables de interes}

\item{window}{Especificar distancia temporal entre las observaciones. anual o trimestral}
}
\value{
Devuelve el pool de datos de panel
}
\description{
Permite armar un pool de datos en panel de la EPH continua a partir
de especificar una serie consecutiva de bases, variables y el largo de la ventana -window- de observación
}
\details{
Disclaimer: El script no es un producto oficial de INDEC.
}
\examples{

lista_bases <- list(toybase_individual_2016_03,toybase_individual_2016_04)
pool_trimestral <- organize_panels(bases = lista_bases,
                    variables =c("P21","ESTADO"),
                    window = "trimestral")
}
