% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organize_rama.R
\name{organize_rama}
\alias{organize_rama}
\title{organize_rama}
\usage{
organize_rama(x)
}
\arguments{
\item{x}{un integer que es la variable PP04B_COD}
}
\description{
Funcion que crea la variable rama de actividad, a partir de la variablo PP04B_COD
}
\details{
Crea variable Rama de Actividad
}
\examples{
df <- get_microdata(year = 2018, trimester = 4, type = "individual")
df$rama.eph <- organize_rama(x = df$PP04B_COD)
df <- df \%>\% dplyr::mutate(rama.eph = organize_rama(x = PP04B_COD))
}
