% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orgnize_panels.R
\name{orgnize_panels}
\alias{orgnize_panels}
\title{Pool de Datos en Panel - Base Individudal EPH continua}
\usage{
orgnize_panels(bases, variables, window = "anual")
}
\arguments{
\item{bases}{Lista de bases de microdatos a utilizar para armar el pool de datos}

\item{variables}{Vector con nombres de las variables de interés}

\item{window}{Especificar distancia temporal entre las observaciones. anual o trimestral}
}
\value{
Devuelve el pool de datos de panel
}
\description{
Permite armar un pool de datos en panel de la EPH continua a partir
de especificar una serie consecutiva de bases, variables y el largo de la window de observación
}
\details{
disclaimer: El script no es un producto oficial de INDEC.
}
\examples{
base_1t_2018 <- get_microdata(year = 2018,trimester = 1)[[2]]
base_2t_2018 <- get_microdata(year = 2018,trimester = 2)[[2]]

lista_bases <- list(base_1t_2018,base_2t_2018)
pool_trimestral <- orgnize_panels(bases = lista_bases,
                    variables =c("P21","ESTADO"),
                    window = "trimestral")
}
