% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/put_labels_eph.R
\name{put_labels_eph}
\alias{put_labels_eph}
\title{Función para etiquetar las bases de la Encuesta Permanente de Hogares.}
\usage{
put_labels_eph(df, base = "individual")
}
\arguments{
\item{df:}{base de microdatos de la EPH
base: string. aclaracion de si se quiere etiquetar una base 'individual' u 'hogar'}
}
\description{
Función para etiquetar las bases de la Encuesta Permanente de Hogares.
}
\details{
disclaimer: El script no es un producto oficial de INDEC.
}
\examples{
df <- get_bases_eph(anio = 2018, trimestre = 1, etiqueta = FALSE)
df <- put_labels_eph(df$base_hogar, base='hogar')
}
