% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagger_eph.R
\name{tagger_eph}
\alias{tagger_eph}
\title{Función para etiquetar las bases de la Encuesta Permanente de Hogares.}
\usage{
tagger_eph(df, base = "individual")
}
\arguments{
\item{df:}{base de microdatos de la EPH
base: string. aclaracion de si se quiere etiquetar una base 'individual' u 'hogar'}
}
\description{
Función para etiquetar las bases de la Encuesta Permanente de Hogares.
}
\details{
disclaimer: El script no es un producto oficial de INDEC.
}
\examples{
df <- descarga_base_eph(anio = 2018, trimestre = 1, etiqueta = FALSE)[['hogar']]
df <- tagger_eph(df, base='hogar')
}
