% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toybase_individual_2016_03.R
\docType{data}
\name{toybase_individual_2016_03}
\alias{toybase_individual_2016_03}
\title{Seleccion aleatoria de casos de la base 2016 trimestre 3 para la base individual}
\format{
Un data frame con 2000 filas y 177 variables

\describe{
\item{CODUSU}{'Codigo para distinguir VIVIENDAS, permite aparearlas con Hogares y Personas. Ademas permite hacer el seguimiento a traves de los trimestres'}
\item{NRO_HOGAR}{' Codigo para distinguir HOGARES, permite aparearlos con Personas'}
\item{COMPONENTE}{'Numero de componente: Nro de orden que se asigna a las personas que conforman cada hogar de la vivienda'}
\item{H15}{'Entrevista individual realizada'}
\item{ANO4}{'Anio de relevamiento'}
\item{TRIMESTRE}{'Ventana de Observacion'}
\item{REGION}{'Codigo de Region'}
\item{MAS_500}{'Aglomerados segun tamanio'}
\item{AGLOMERADO}{'Codigo de Aglomerado'}
\item{PONDERA}{'Ponderacion'}
\item{CH03}{'Relacion de parentesco'}
\item{CH04}{'Sexo'}
\item{CH05}{'Fecha de nacimiento (dia, mes y anio)'}
\item{CH06}{'Edad en anios cumplidos'}
\item{CH07}{'Actuamente esta...'}
\item{CH08}{'Tipo de cobertura medica'}
\item{CH09}{'Sabe leer y escribir?'}
\item{CH10}{'Asiste o asistio a algun establecimiento educativo? (colegio, escuela, universidad)'}
\item{CH11}{'Ese establecimiento es...'}
\item{CH12}{'Cual es el nivel mas alto que cursa o curso?'}
\item{CH13}{'Finalizo ese nivel?'}
\item{CH14}{'Cual fue el ultimo anio que aprobo?'}
\item{CH15}{'Donde nacio?'}
\item{CH15_COD}{'Codigo de provincia / pais / territorio'}
\item{CH16}{'Donde vivia hace 5 anios?'}
\item{CH16_COD}{'Especificar'}
\item{NIVEL_ED}{'Nivel Educativo'}
\item{ESTADO}{'Condicion de actividad'}
\item{CAT_OCUP}{'Categoria ocupacional'}
\item{CAT_INAC}{'Categoria de inactividad'}
\item{IMPUTA}{'Indica los casos que han sido imputados'}
\item{PP02C1}{'De Que manera estuvo buscando trabajo? - Hizo contactos, entrevistas'}
\item{PP02C2}{'De Que manera estuvo buscando trabajo? - Mando curriculum, puso o contesto avisos'}
\item{PP02C3}{'De Que manera estuvo buscando trabajo? - Se presento en establecimientos'}
\item{PP02C4}{'De Que manera estuvo buscando trabajo? - Hizo algo para ponerse por su cuenta'}
\item{PP02C5}{'De Que manera estuvo buscando trabajo? - Puso carteles en negocios, pregunto en el barrio'}
\item{PP02C6}{'De Que manera estuvo buscando trabajo? - Consulto a parientes, amigos'}
\item{PP02C7}{'De Que manera estuvo buscando trabajo? - Se anioto en bolsas, listas, planes de empleo, agencias, contratistas o alguien le estpa buscando trabajo'}
\item{PP02C8}{'De Que manera estuvo buscando trabajo? - Busco trabajo de otra forma activa'}
\item{PP02E}{'Durante esos 30 dias no busCo trabajo, porque...'}
\item{PP02H}{'En los ultimos 12 meses, busco trabajo en algun momento?'}
\item{PP02I}{'En los ultimos 12 meses, trabajo en algun momento?'}
\item{PP03C}{'La semana pasada, tenia...'}
\item{PP03D}{'Cantidad de ocupaciones.'}
\item{PP3E_TOT}{'Total de horas que trabajo en la semana en la ocupacion principal'}
\item{PP3F_TOT}{'Total de horas que trabajo en la semana en otras ocupaciones'}
\item{PP03G}{'La semana pasada, queria trabajar mas horas?'}
\item{PP03H}{'Si hubiera conseguido mas horas...'}
\item{PP03I}{'En los ultimos 30 dias, busco trabajar mas horas?'}
\item{PP03J}{'Aparte de este/os trabajo/s, estuvo buscando algun empleo / ocupacion / actividad?'}
\item{INTENSI}{'INTENSI'}
\item{PP04A}{'El negocio / empresa / institucion / actividad en la que trabaja es (se refiere al que trabaja mas horas semanales)...'}
\item{PP04B_COD}{'A que se dedica o produce el negocio / empresa / institucion?'}
\item{PP04B1}{'Si presta servicio domestico en hogares particulares'}
\item{PP04B2}{'En cuantas casas trabaja?'}
\item{PP04B3_MES}{'Cuanto tiempo hace que trabaja alli? - MES'}
\item{PP04B3_ANO}{'Cuanto tiempo hace que trabaja alli? - ANIO'}
\item{PP04B3_DIA}{'Cuanto tiempo hace que trabaja alli? - DiAS'}
\item{PP04C}{'Cuantas personas, incluido... trabajan alli en total?'}
\item{PP04C99}{'Tamanio del establecimiento'}
\item{PP04D_COD}{ 'Codigo de ocupacion'}
\item{PP04G}{'Donde realiza principalmente sus tareas?'}
\item{PP05B2_MES}{'Cuanto tiempo hace que trabaja en ese emPPleo en forma continua? - MES'}
\item{PP05B2_ANO}{'Cuanto tiempo hace que trabaja en ese emPPleo en forma continua? - ANIO'}
\item{PP05B2_DIA}{'Cuanto tiempo hace que trabaja en ese emPPleo en forma continua? - DIA'}
\item{PP05C_1}{'En ese negocio / empresa / actividad, tiene... maquinarias / equipos?'}
\item{PP05C_2}{'En ese negocio / empresa / actividad, tiene... local (incluye kiosko, puesto fijo)?'}
\item{PP05C_3}{'En ese negocio / empresa / actividad, tiene... vehiculo?'}
\item{PP05E}{'Para la actividad del negocio, en los ultimos 3 meses, tuvo que gastar en la compra de materias primas, productos, pagar servicios u otros gastos?'}
\item{PP05F}{'Ese negocio / empresa / actividad, trabaja habitualmente para...'}
\item{PP05H}{'Durante cuanto tiempo ha estado trabajando en ese empleo de forma continua? (con interrupciones laborales no mayores a 15 dias)'}
\item{PP06A}{'En ese negocio / empresa / actividad tiene socios o familiares asociados?'}
\item{PP06C}{'Monto de patrones y cuenta propia sin socios'}
\item{PP06D}{'Monto de patrones y cuenta propia con socios'}
\item{PP06E}{'Ese negocio / empresa / actividad...'}
\item{PP06H}{'Es una actividad familiar?'}
\item{PP07A}{'Cuanto tiempo hace que esta trabajando en ese empleo en forma continua? (sin interrupccion de la relacion laboral en la misma empresa / negocio / institucion)'}
\item{PP07C}{'Ese empleo tiene tiempo de finalizacion?'}
\item{PP07D}{'Por cuanto tiempo es ese trabajo?'}
\item{PP07E}{'Ese trabajo es...?'}
\item{PP07F1}{'En ese trabajo le dan... de comer gratis en el lugar de trabajo'}
\item{PP07F2}{'En ese trabajo le dan... vivienda?'}
\item{PP07F3}{'En ese trabajo le dan... algun producto o mercaderia?'}
\item{PP07F4}{'En ese trabajo le dan... aglun otro beneficio? (automovil, telefono celular, pasajes, etc.)'}
\item{PP07F5}{'En ese trabajo le dan... no recibe ninguno?'}
\item{PP07G1}{'En ese trabajo tiene... vacaciones pagas?'}
\item{PP07G2}{'En ese trabajo tiene... aguinaldo?'}
\item{PP07G3}{'En ese trabajo tiene... dias pagos por enfermedad?'}
\item{PP07G4}{'En ese trabajo tiene... obra social?'}
\item{PP07G_59}{'En ese trabajo tiene... no tiene ninguno?'}
\item{PP07H}{'Por ese trabajo tiene descuento jubilatorio?'}
\item{PP07I}{'Aporta por si mismo a algun sistema jubilatorio?'}
\item{PP07J}{'El turno habitual de trabajo es...?'}
\item{PP07K}{'Cuando cobra...'}
\item{PP08D1}{'Cuanto cobro por ese mes por esos conceptos? Monto total de sueldos/jornales, salario familiar, horas extras, otras bonificaciones habituales y tickets, vales o similares'}
\item{PP08D4}{'Por el mes de...(mes)....cobro... Monto percibido en tickets.'}
\item{PP08F1}{'Cuanto cobro por ese mes de ..(mes)... Monto en pesos cobrado por Comision por venta/produccion.'}
\item{PP08F2}{'Cuanto cobro por ese mes de ..(mes)... Monto en pesos cobrado por Propinas.'}
\item{PP08J1}{'Cuanto cobro por ese mes de ..(mes)... Monto aguinaldo.'}
\item{PP08J2}{'Cuanto cobro por ese mes de ..(mes)... Monto otras bonificaciones no habituales.'}
\item{PP08J3}{'Cuanto cobro por ese mes de ..(mes)... Monto retroactivos.'}
\item{PP09A}{'SOLO OCUPADOS DE: CIUDAD DE BUENOS AIRES.  Partidos del GBA En su ocupacion (o en la de mas horas)trabaja....en ....'}
\item{PP09A_ESP}{'Especificar: contiene la descripcion de otro lugar.'}
\item{PP09B}{'En su ocupacion (o en la de mas horas), trabaja... en esta ciudad?'}
\item{PP09C}{'Donde trabaja...?'}
\item{PP09C_ESP}{'Descripcion de otro lugar (segun pregunta PP09C).'}
\item{PP10A}{'Cuanto hace que esta buscando trabajo..?'}
\item{PP10C}{'Durante ese tiempo hizo algun trabajo / changa'}
\item{PP10D}{'Has trabajdo alguna vez?'}
\item{PP10E}{'Cuanto tiempo hace que termino su ultimo trabajo / changa?'}
\item{PP11A}{'El negocio / empresa / institucion / actividad en la que trabaja era...?'}
\item{PP11B_COD}{'A que se dedicaba o que producia ese negocio/empresa/institucion?'}
\item{PP11B1}{'Si prestaba servicios domesticos en hogares particulares,'}
\item{PP11B2_DIA}{'Cuanto tiempo trabajo alli? - DIAS'}
\item{PP11B2_MES}{'Cuanto tiempo trabajo alli? - MESES'}
\item{PP11B2_ANO}{'Cuanto tiempo trabajo alli? - ANIOS'}
\item{PP11C}{'Cuantas personas, inluido ... trabajan alli en total?'}
\item{PP11C99}{'Ns./Nr. cuantas personas trabajan alli en total'}
\item{PP11D_COD}{'Como se llamaba la ocupacion que tenia?'}
\item{PP11G_ANO}{'Cuanto tiempo seguido estuvo trabajando en ese lugar? - ANIO'}
\item{PP11G_MES}{'Cuanto tiempo seguido estuvo trabajando en ese lugar? - MES'}
\item{PP11G_DIA}{'Cuanto tiempo seguido estuvo trabajando en ese lugar? - DIA'}
\item{PP11L}{'Cual fue la razon principal por la que dejo esa actividad?'}
\item{PP11L1}{'Ese trabajo era...'}
\item{PP11M}{'Ese trabajo era...'}
\item{PP11N}{'En ese trabajo lehacian descuento jubilatorio?'}
\item{PP11O}{'Cual fue la razon principal por la que dejo ese trabajo?'}
\item{PP11P}{'Cerro la empresa?'}
\item{PP11Q}{ 'Fue la unica persona que quedo sin trabajo?'}
\item{PP11R}{'Le enviaron telegrama?'}
\item{PP11S}{'Le pagaron indemnizacion?'}
\item{PP11T}{'Esta cobrando seguro de desempleo?'}
\item{P21}{'MONTO DE INGRESO DE LA OCUPACIoN PRINCIPAL.'}
\item{DECOCUR}{'Nro de decil de ingreso de la ocupacion principal del TOTAL EPH.'}
\item{IDECOCUR}{'Nro de decil de ingreso de la ocupacion principal del INTERIOR EPH'}
\item{RDECOCUR}{'Nro de decil de ingreso de la ocupacion principal de la REGION'}
\item{GDECOCUR}{'Nro de decil de ingreso de la ocupacion principal del CONJUNTO DE AGLOMERADOS DE 500 MIL Y MAS HABITANTES'}
\item{PDECOCUR}{'Nro de decil de ingreso de la ocupacion principal del CONJUNTO DE AGLOMERADOS DE MENOS DE 500 MIL HABITANTES'}
\item{ADECOCUR}{'Nro de decil de ingreso de la ocupacion principal del AGLOMERADO.'}
\item{PONDIIO}{'Ponderador del ingreso de la ocupacion principal.'}
\item{TOT_P12}{'MONTO DE INGRESO DE OTRAS OCUPACIONES. (Incluye: ocupacion secundaria, ocupacion previa a la semana de referencia, deudas/retroactivos por ocupaciones anteriores al mes de referencia,}  etc).',
\item{P47T}{'MONTO DE INGRESO TOTAL INDIVIDUAL(sumatoria ingresos laborales y no laborales).'}
\item{DECINDR}{'Nro de decil de ingreso total individual del TOTAL EPH.'}
\item{IDECINDR}{'Nro de decil de ingreso total individual del INTERIOR EPH.'}
\item{RDECINDR}{'Nro de decil de ingreso total individual de la REGION.'}
\item{GDECINDR}{'Nro de decil de ingreso total individual del CONJUNTO DE AGLOMERADOS DE 500 MIL Y MAS HABITANTES.'}
\item{PDECINDR}{'Nro de decil de ingreso total individual del CONJUNTO DE AGLOMERADOS DE MENOS DE 500 MIL HABITANTES.'}
\item{ADECINDR}{'Nro de decil de ingreso total individual del AGLOMERADO.'}
\item{PONDII}{'Ponderador para ingreso total individual.'}
\item{V2_M}{'Monto del ingreso por JUBILACION O PENSION.'}
\item{V3_M}{'Monto del ingreso por INDEMNIZACION POR DESPIDO.'}
\item{V4_M}{'Monto del ingreso por SEGURO DE DESEMPLEO.'}
\item{V5_M}{'Monto del ingreso por SUBSIDIO O AYUDA SOCIAL (EN DINERO) DEL GOBIERNO, IGLESIAS, ETC.'}
\item{V8_M}{'Monto del ingreso por ALQUILER (VIVIENDA , TERRENO, OFICINA, ETC.) DE SU PROPIEDAD.'}
\item{V9_M}{'Monto del ingreso por GANANCIAS DE ALGUN NEGOCIO EN EL QUE NO TRABAJo.'}
\item{V10_M}{'Monto del ingreso por INTERESES O RENTAS POR PLAZOS FIJOS/INVERSIONES.'}
\item{V11_M}{'Monto del ingreso por BECA DE ESTUDIO.'}
\item{V12_M}{'Monto del ingreso por CUOTAS DE ALIMENTOS O AYUDA EN DINERO DE PERSONAS QUE NO VIVEN EN EL HOGAR.'}
\item{V18_M}{'Monto del ingreso por OTROS INGRESOS EN EFECTIVO (LIMOSNAS, JUEGOS DE AZAR, ETC.).'}
\item{V19_AM}{'Monto del ingreso por TRABAJO DE MENORES DE 10 ANIOS.'}
\item{V21_M}{'Monto del ingreso por aguinaldo.'}
\item{T_VI}{'MONTO TOTAL DE INGRESOS NO LABORALES'}
\item{ITF}{'MONTO DEL INGRESO TOTAL FAMILIAR.'}
\item{DECIFR}{'Nro de decil de ingreso total familiar del TOTAL.'}
\item{IDECIFR}{'Nro de decil de ingreso total familiar del INTERIOR.'}
\item{RDECIFR}{'Nro de decil de ingreso total familiar de la REGION.'}
\item{GDECIFR}{'Nro de decil de ingreso total familiar del CONJUNTO DE AGLOMERADOS DE 500 MIL Y MAS HABITANTES.'}
\item{PDECIFR}{'Nro de decil de ingreso total familiar del CONJUNTO DE AGLOMERADOS DE MENOS DE 500 MIL HABITANTES.'}
\item{ADECIFR}{'Nro de decil de ingreso total familiar del AGLOMERADO (ver Anexo I).'}
\item{IPCF}{'MONTO DEL INGRESO PER CAPITA FAMILIAR.'}
\item{DECCFR}{'Nro de decil de ingreso per capita familiar del TOTAL EPH (ver Anexo I).'}
\item{IDECCFR}{'Nro de decil de ingreso per capita familiar del  INTERIOR EPH (ver Anexo I).'}
\item{RDECCFR}{'Nro de decil de ingreso per capita familiar de la REGION.'}
\item{GDECCFR}{'Nro de decil de ingreso per capita familiar del CONJUNTO DE AGLOMERADOS DE 500 MIL Y MAS HABITANTES.'}
\item{PDECCFR}{'Nro de decil de ingreso per capita familiar del CONJUNTO DE AGLOMERADOS DE MENOS DE 500 MIL HABITANTES.'}
\item{ADECCFR}{'Nro de decil de ingreso per capita familiar del AGLOMERADO.'}
\item{PONDIH}{'Ponderador del ingreso total familiar y del ingreso per capita familiar, para hogares.}
...
}
}
\usage{
toybase_individual_2016_03
}
\description{
Seleccion aleatoria de casos de la base 2016 trimestre 3 para la base individual
}
\keyword{datasets}
