% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_poverty.R
\name{calculate_poverty}
\alias{calculate_poverty}
\title{Calculo de Pobreza e Indigencia}
\usage{
calculate_poverty(base, basket, print_summary = TRUE)
}
\arguments{
\item{base}{Base individual de uno o mas periodos}

\item{basket}{canasta basica alimentaria y total, con la estructura de la canasta de ejemplo (ver \code{\link{canastas_reg_example}})}

\item{print_summary}{default = TRUE. Opcion para imprimir las tasas de pobreza e indigencia (proporcion de individuos)}
}
\description{
Funcion para calcular la pobreza e indigencia siguiendo la metodologia de linea.
}
\details{
disclaimer: El script no es un producto oficial de INDEC.
}
\examples{

bases <- dplyr::bind_rows(toybase_individual_2016_03,
                          toybase_individual_2016_04)

base_pobreza <- calculate_poverty(base = bases,
                                  basket = canastas_reg_example,
                                  print_summary=TRUE)

}
