% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_tabulates.R
\name{calculate_tabulates}
\alias{calculate_tabulates}
\title{Tabulado con ponderación}
\usage{
calculate_tabulates(base, x, y = NULL, weights = NULL, digits = 3,
  normwt = FALSE, na.rm = TRUE, na.show = FALSE, exclude = NULL,
  add.totals = "none", add.percentage = "none")
}
\arguments{
\item{base}{Dataframe}

\item{x}{string con el nombre de la variable a tabular}

\item{y}{otro string (opcional) con el nombre de una segunda variable, para una tabla de doble entrada. Tiene que ser de igual largo que x}

\item{weights}{string con el nombre de la variable con los pesos de pesos, tiene que ser de igual largo que x}

\item{digits}{numero de digitos significativos}

\item{normwt}{si es TRUE, normaliza los pesos de modo que el recuento ponderado total sea el mismo que el no ponderado}

\item{na.rm}{Si es TRUE, elimina los NA antes del computo}

\item{na.show}{si TRUE, muestra el recuento de NA en la salida}

\item{exclude}{valores a remover de x e y}

\item{add.totals}{toma los valores c('none','row','col','both'), para agregar totales por fila, columna o ambos}

\item{add.percentage}{toma los valores c('none','row','col'), para agregar porcentajes por fila y columna}
}
\description{
Función para crear tabulados uni o bivariados con ponderacion, totales parciales y porcentajes.
}
\examples{


### tabla simple ###

calculate_tabulates(base = toybase_individual_2016_04,x = 'REGION', y = 'CH04',
            weights = 'PONDERA')

### para ver los totales por fila ###

calculate_tabulates(base = toybase_individual_2016_04,x = 'REGION', y = 'CH04',
            weights = 'PONDERA', add.totals = 'row')

### para ver los totales por columna ###

calculate_tabulates(base = toybase_individual_2016_04,x = 'REGION', y = 'CH04',
            weights = 'PONDERA', add.totals = 'col')

### para ver porcentajes por fila ###

calculate_tabulates(base = toybase_individual_2016_04,x = 'REGION', y = 'CH04',
            weights = 'PONDERA', add.percentage ='row')

}
