% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canastas_reg_example.R
\docType{data}
\name{canastas_reg_example}
\alias{canastas_reg_example}
\title{Canastas Básicas Alimentarias y Cavnastas Básicas Totales según región y trimestre}
\format{A data frame with 12 rows and 5 variables

\describe{
\item{region}{Nombre de región}
\item{periodo}{Año.trimestre}
\item{CBA}{canasta basica alimentaria}
\item{CBT}{inversa del coeficiente de engels}
\item{codigo}{código de región}
...
}}
\usage{
canastas_reg_example
}
\description{
Esta base tiene el formato necesario para el calculo de pobreza
}
\keyword{datasets}
