% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNO.R
\docType{data}
\name{CNO}
\alias{CNO}
\title{Categorias de las 4 dimensiones del Clasificador Nacional de Ocupaciones 2001.}
\format{A data frame with 63 rows and 4 variables

\describe{
\item{digit}{digito para identificar la dimensión del CNO}
\item{value}{valor que toma cada categoría dentro de una dimensión}
\item{label}{nombre de la categoría propia a una dimensión}
\item{variable}{nombre de la dimensión del CNO}
...
}}
\usage{
CNO
}
\description{
Categorias de las 4 dimensiones del Clasificador Nacional de Ocupaciones 2001.
}
\keyword{datasets}
