% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descarga_base_eph.r
\name{descarga_base_eph}
\alias{descarga_base_eph}
\title{Función que descarga bases de la Encuesta Permanente de Hogares del INDEC.}
\usage{
descarga_base_eph(anio = 2018, trimestre = 1, etiqueta = FALSE)
}
\arguments{
\item{anio:}{un integer entre 2016:2018
trimestr: un integer con el numero de trimestre: 1,2,3,4
etiqueta: TRUE/FALSE, opcion para etiquetar los datos}
}
\description{
Función que descarga bases de la Encuesta Permanente de Hogares del INDEC.
}
\details{
disclaimer: El script no es un producto oficial de INDEC.
}
