% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diccionario_regiones.R
\docType{data}
\name{diccionario_regiones}
\alias{diccionario_regiones}
\title{Diccionario de regiones segun diseño de registro de la EPH}
\format{
A data frame with 6 rows and 2 variables

\describe{
\item{codigo}{Codigo de region}
\item{region}{Nombre de región}

}
}
\usage{
diccionario_regiones
}
\description{
Diccionario de regiones segun diseño de registro de la EPH
}
\keyword{datasets}
