% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diccionario_regiones.R
\docType{data}
\name{diccionario_regiones}
\alias{diccionario_regiones}
\title{Diccionario de regiones segun diseno de registro de la EPH}
\format{
Un data frame con 6 filas y 2 variables

\describe{
\item{codigo}{Codigo de region}
\item{region}{Nombre de region}

}
}
\usage{
diccionario_regiones
}
\description{
Diccionario de regiones segun diseno de registro de la EPH
}
\keyword{datasets}
