% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eph.R
\docType{package}
\name{eph}
\alias{eph}
\title{\code{eph} package}
\description{
Caja de Herramientas para el procesamiento de la Encuesta Permanente de Hogares
}
\details{
See the README on
\href{https://github.com/pablinte/eph/blob/master/README.md}{Github}
}
