% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_poverty_lines.R
\name{get_poverty_lines}
\alias{get_poverty_lines}
\title{Descarga de canasta basica alimentaria y canasta basica total}
\usage{
get_poverty_lines()
}
\description{
Función que descarga la CBA y CBT a partir de abril 2016
}
\details{
disclaimer: El script no es un producto oficial de INDEC.

La función descarga la serie provista por indec en https://www.indec.gob.ar/indec/web/Nivel4-Tema-4-43-149
}
\examples{

canasta <- get_poverty_lines()

}
