% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_poverty_lines.R
\name{get_poverty_lines}
\alias{get_poverty_lines}
\title{Descarga de canasta basica alimentaria y canasta basica total}
\usage{
get_poverty_lines(regional = FALSE)
}
\arguments{
\item{regional}{booleano, default = FALSE. Si es TRUE, descarga los datos de canastas regionales
que se utilizan para el calculo de pobreza. Si es FALSE, descarga la serie de GBA
provista por indec en https://www.indec.gob.ar/indec/web/Nivel4-Tema-4-43-149}
}
\value{
Devuelve una tabla con la CBA y CBT a partir de abril 2016
}
\description{
Funcion que descarga la CBA y CBT a partir de abril 2016
}
\details{
disclaimer: El script no es un producto oficial de INDEC.
}
\examples{

canasta <- get_poverty_lines(regional = TRUE)
canasta

}
