% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organize_caes.R
\name{organize_caes}
\alias{organize_caes}
\title{organize_caes}
\usage{
organize_caes(base)
}
\arguments{
\item{base}{Base individual de uno o mas periodos}
}
\value{
Devuelve la base con 8 columnas nuevas (ver \code{\link{caes}})
}
\description{
Funcion para clasificar las actividades economicas segun el
             Clasificador de actividades economicas para encuestas sociodemograficas.
             CAES Mercosur 1.0 y CAES Mercosur.
             Basado en https://www.indec.gob.ar/ftp/cuadros/menusuperior/clasificadores/notas_explicativas_caes_v2018.pdf
}
\details{
disclaimer: El script no es un producto oficial de INDEC.

El usuario debe contemplar los cambios en la codificacion y categorias de las distintas versiones del clasificador CAES, para la elaboracion de series de largo plazo.
}
\examples{
bases <- dplyr::bind_rows(toybase_individual_2016_03,toybase_individual_2016_04)
organize_caes(base = bases)

}
