% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organize_labels.R
\name{organize_labels}
\alias{organize_labels}
\title{Funcion para etiquetar las bases de la Encuesta Permanente de Hogares.}
\usage{
organize_labels(df, type = "individual")
}
\arguments{
\item{df}{type de microdatos de la EPH}

\item{type}{(string) aclaracion sobre si la base a etiquetar es de tipo 'individual' u 'hogar'}
}
\description{
Funcion para etiquetar las bases de la Encuesta Permanente de Hogares.
}
\details{
Disclaimer: El script no es un producto oficial de INDEC.
}
\examples{
df <- organize_labels(toybase_individual_2016_04, type='individual')
}
