% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pool_datapanel_eph.R
\name{pool_datapanel_eph}
\alias{pool_datapanel_eph}
\title{Pool de Datos en Panel - Base Individudal EPH continua}
\usage{
pool_datapanel_eph(bases, variables, ventana = "anual")
}
\arguments{
\item{bases}{Lista de bases de microdatos a utilizar para armar el pool de datos}

\item{variables}{Vector con nombres de las variables de interés}

\item{ventana}{Que tipo de distancia temporal entre las obseraciones. anual o trimestral}
}
\value{
Devuelve el pool de datos de panel y un conteo de casos consistentes
}
\description{
Permite armar un pool de datos en panel de la EPH continua a partir
de especificar una serie de bases, variables y el largo de la ventana de observación
}
\examples{
base_1t_2018 <- get_bases_eph(anio = 2018,trimestre = 1)[[2]]
base_2t_2018 <- get_bases_eph(anio = 2018,trimestre = 2)[[2]]

lista_bases <- list(base_1t_2018,base_2t_2018)
pool_datapanel_eph(bases = lista_bases,
                    variables =c("P21","ESTADO"),
                    ventana = "trimestral")
}
