% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtd_crosstab.R
\name{wtd_crosstab}
\alias{wtd_crosstab}
\title{Tabulado bi-variado con ponderación}
\usage{
wtd_crosstab(base, var1, var2, weight_var, total = "col", na.rm = TRUE)
}
\arguments{
\item{base:}{data.frame donde se alojan las variables.}

\item{var1:}{Variable en fila.}

\item{var2:}{Variable en columnas.}

\item{weight_var:}{Variable correspondiente al factor de ponderación.}

\item{na.rm:}{TRUE/FALSE, en caso de quere/eliminar los valores NA.}

\item{pct_col:}{TRUE/FALSE si el 100% es por columnas.}

\item{pct_row:}{TRUE/FALSE si el 100% es por filas.}
}
\description{
Función que genera un cruce de variables ponderadas por una tercera.
}
\details{
disclaimer: El script no es un producto oficial de INDEC.
}
\examples{
wtd.crosstab(var1 = df$var1, var2 = df$var2, weight_var = df$weighted_var, total = "col", na.rm=TRUE)
}
